kvdb-cli prompt delete confirmation

+++ cmd/kvdb-cli/cmd/db/delete.go
@@ -3,6 +3,7 @@ package db
 import (
 	"context"
 	"fmt"
+	"strings"
 
 	"github.com/hollowdll/kvdb/cmd/kvdb-cli/client"
 	"github.com/hollowdll/kvdb/cmd/kvdb-cli/config"
@@ -26,14 +27,31 @@ func init() {
 }
 
 func deleteDatabase() {
+	if !promptConfirmDelete() {
+		return
+	}
 	ctx := metadata.NewOutgoingContext(context.Background(), client.GetBaseGrpcMetadata())
 	ctx, cancel := context.WithTimeout(ctx, client.CtxTimeout)
 	defer cancel()
 	if len(dbName) < 1 {
 		dbName = viper.GetString(config.ConfigKeyDatabase)
 	}
+
 	res, err := client.GrpcDatabaseClient.DeleteDatabase(ctx, &kvdbserver.DeleteDatabaseRequest{DbName: dbName})
 	client.CheckGrpcError(err)
 
 	fmt.Println(res.DbName)
+}
+
+func promptConfirmDelete() bool {
+	var input string
+	fmt.Print("Delete database and all its data? Yes/No: ")
+	_, err := fmt.Scanln(&input)
+	input = strings.TrimSpace(input)
+	if input == "" {
+		return false
+	}
+	cobra.CheckErr(err)
+
+	return strings.ToLower(input) == "yes"
 }
